gpu: nvgpu: pmu: cast assignments of sizeof to u32
This change fixes a number of This is a MISRA 10.3 rule violation due to
the implicit casts of sizeof() to u32's. This change adds u32 casts to
each of these violations. This should be safe because a 4GB type size
would be very unlikely in this driver.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/1959429
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/pmu/pmu.c
@@ -385,7 +385,7 @@ int nvgpu_pmu_process_init_msg(struct nvgpu_pmu *pmu,
 	}
 
 	nvgpu_falcon_copy_from_dmem(pmu->flcn, tail + PMU_MSG_HDR_SIZE,
-		(u8 *)&msg->msg, msg->hdr.size - PMU_MSG_HDR_SIZE, 0);
+		(u8 *)&msg->msg, (u32)msg->hdr.size - PMU_MSG_HDR_SIZE, 0);
 
 	if (msg->msg.init.msg_type != PMU_INIT_MSG_TYPE_PMU_INIT) {
 		nvgpu_err(g, "expecting init msg");
@@ -402,7 +402,7 @@ int nvgpu_pmu_process_init_msg(struct nvgpu_pmu *pmu,
 		nvgpu_falcon_copy_from_dmem(pmu->flcn,
 			pv->get_pmu_init_msg_pmu_sw_mg_off(init),
 			(u8 *)&gid_data,
-			sizeof(struct pmu_sha1_gid_data), 0);
+			(u32)sizeof(struct pmu_sha1_gid_data), 0);
 
 		pmu->gid_info.valid =
 			(*gid_hdr_data == PMU_SHA1_GID_SIGNATURE);
