+++ drivers/gpu/nvgpu/common/pmu/pmu_gp106.c
@@ -171,7 +171,7 @@ void gp106_pmu_elpg_statistics(struct gk20a *g, u32 pg_engine_id,
 
 	nvgpu_falcon_copy_from_dmem(pmu->flcn,
 		pmu->stat_dmem_offset[pg_engine_id],
-		(u8 *)&stats, sizeof(struct pmu_pg_stats_v2), 0);
+		(u8 *)&stats, (u32)sizeof(struct pmu_pg_stats_v2), 0);
 
 	pg_stat_data->ingating_time = stats.total_sleep_time_us;
 	pg_stat_data->ungating_time = stats.total_non_sleep_time_us;
